Add 'Export All Jobs' functionality
Functionality to export all ongoing jobs at once has been implemented, allowing users to specify a directory to export to in-bulk. While this works, it has been hard-coded to export jobs as .PNG images when this option is used for the time being, and a solution will need to be implemented for allowing users to specify a format of their choosing in the future.