Introduced nlohmann's JSON as a project dependency
Niels Lohmann's JSON for Modern C++ library has been introduced as a dependency; with the 'single include' version of the file being copied to an appropraite place within `src/`.

This has been done in order to facilitate the introduction of JSON support into doomify and support for JSON files as a type of palette that can be imported by users.

An attribution notice has also been added to the README and a copy of the MIT License included within the directory holding the library's files to ensure compliance with its terms.