Created PaletteManager class with PullAllPalettesFromDisk method.
A new PaletteManager class has been created with an empty PullAllPalettesFromDisk method definition inside it. This has been done in order to begin the refactoring process of how Palettes are loaded in from multiple file types with the intent being this manager acting as an abstraction layer for loading and returning all Palettes found on disk from other dedicated 'Loader' classes based on the file type.