Created base PaletteLoader class and PaletteLoaderOutcome Enum.
A base PaletteLoader class has been created with related virtual functions to define a contract for classes meant for loading palettes from a given file type to obey. Likewise, a PaletteLoaderOutcome enum has been created as part of this with the intent of hopefully facilitating proper error checking when attempting to load in Palettes from an external file.