Moved old GPL parsing logic into GplPaletteLoader
The existing logic for loading GPL palettes within `palettes.cpp` has been moved over to GplPaletteLoader as part of the refactor to seperate logic for parsing different file types as palettes.

To test these changes, and because this would have to be done at some point anyway, logic has also been adjusted in the MainWindow to pull palettes from the PaletteManager which calls GplPaletteLoader by proxy. This has the disadvantage of only GPL files being supported in this current branch, but this will be resolved as existing logic for the other file types are brought over into their own PaletteLoaders.