Added Show Palettes Directory button to UI and connected signal/slot in PaletteManager
The Qt UI file has been updated to introduce a button alongside the palette selection dropdown which is planned to open the user's palettes directory when clicked. A signal and slot have been connected through the MainWindow and PaletteManager respectively for when this button is clicked and executes correctly - though the desired behaviour itself has not been implemented yet and will need to be later on. This work is being done as part of efforts to allow users to load in new palettes during runtime without having to restart the application.